a somewhat functional implementation of a delete function
I can not predict when it will cause the program to crash.  My belief is that the program does not move back through the nodes.  So it gets stuck on some child or sibling node and can no longer search and remove other words.

My other theory is that you can somehow remove everything from a parent node, or child node with leaves, but it won't rotate the nodes in the child or leaves (respectively) causing a branch to be severed from the tree making other functions not operate properly.

However for a lot of test cases it is able to function properly.