Old Archived Attempt with templates
This was an unrealized attempt to use templates to allow for B-Trees to be done under the methodologies of integers.

The actual B-Tree code was not written by me.  It was adapted from the implementation on Programiz.

This proof of concept worked if I remember correctly as the idea of templates was brought to me while asking for help really early on in the development of this project.

It was scrapped in favor of creating code from the ground up that worked with strings primarily instead of using all datatypes.